[ICC-23] Service 코드 수정 및 생성

+++ src/main/java/com/icc/qasker/quiz/dto/request/AnswerRequest.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class AnswerRequest {
-    private Long number;
+    private int number;
     private Long userAnswer;
 }
